[Refactor] 좋아요 계층 구조 변경 (#4)
댓글 좋아요, 게시글 좋아요 별도의 레이아웃 3계층에서 하나의 like 계층으로 변경
