+++ index.js
@@ -1,10 +1,13 @@
+import {likeRouter} from "./src/routes/Like.route";
+
 const express = require("express");
 const dotenv = require("dotenv");
 const cors = require("cors");
 const swaggerUi = require("swagger-ui-express");
 const specs = require("./config/swagger.config");
 import cookieParser from "cookie-parser";
-const { postRouter, postsRouter } = require("./src/routes/post.route");
+import { postRouter } from "./src/routes/post.route";
+import { likeRouter } from "./src/routes/Like.route";
 import { authRouter } from "./src/routes/auth.route.js"; // .js 확장자 추가
 
 //서버 가동
@@ -23,7 +26,8 @@ app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(specs));
 
 app.use(cookieParser());
 
-app.use("/posts", postsRouter);
+app.use("/posts", postRouter);
+app.use("/likes",likeRouter);
 app.use("/auth", authRouter);
 
 app.get("/", (req, res) => {
