+++ src/controllers/Like.controller.js
@@ -1,0 +1,6 @@
+import { postLikeService } from '../services/Like.service';
+import {status} from "../../config/response.status";
+
+export const postLike = async(req,res,next)=>{
+  return res.send(response(status.SUCCESS, await postLikeService(req.params.postId,req.body)));
+}
