+++ src/models/Like.dao.js
@@ -1,17 +1,17 @@
 import { pool } from '../../config/database.js';
-import { likeWhether,addLike,deleteLike } from './postLike.sql.js';
+import { postLikeWhether,postAddLike,postDeleteLike } from './Like.sql.js';
 
-export const userLike = async(user_id,post_id)=>{
+export const postLike = async(user_id,post_id)=>{
   try{
     const postId = parseInt(post_id, 10);
     const userId = parseInt(user_id, 10);
     const conn = await pool.getConnection();
-    const [like] =await pool.query(likeWhether,[postId, userId])
+    const [like] =await pool.query(postLikeWhether,[postId, userId])
     if(like.length === 0){
-      await pool.query(addLike,[postId, userId])
+      await pool.query(postAddLike,[postId, userId])
       console.log('좋아요 추가');
     }else {
-    await pool.query(deleteLike,[postId, userId])
+    await pool.query(postDeleteLike,[postId, userId])
     console.log('좋아요 취소');
     }
     conn.release(); // 연결 해제
