+++ src/services/post.service.js
@@ -2,8 +2,6 @@ import { addPost } from '../models/post.dao'
 
 export const addPostInfo =async (body) =>{
   const { userId, category, title, image, type, content} = body;
-  console.log(userId)
-
   const result = await addPost(userId, category, title, image, type, content);
   return result
 }
